fix conflect between edit and attach actions
