+++ packages/testsweets/lib/src/enums/capture_widget_enum.dart
@@ -3,5 +3,5 @@ enum CaptureWidgetStatusEnum {
   createWidget,
   editWidget,
   attachWidget,
-  popupMenuShown
+  quickPositionEdit,
 }
