+++ packages/testsweets/lib/src/ui/shared/popup_menu/custom_popup_menu.dart
@@ -51,8 +51,8 @@ class CustomPopupMenu extends StatefulWidget {
     this.position,
     this.menuOnChange,
     required this.onLongPressMoveUpdate,
-    required this.onLongPressUpWhilePopupHidden,
-    required this.onLongPressDown,
+    required this.onLongPressUp,
+    required this.onMoveStart,
     this.onTap,
     this.enablePassEvent = true,
   });
@@ -70,8 +70,8 @@ class CustomPopupMenu extends StatefulWidget {
   final PreferredPosition? position;
   final void Function(bool)? menuOnChange;
   final void Function(LongPressMoveUpdateDetails) onLongPressMoveUpdate;
-  final void Function() onLongPressUpWhilePopupHidden;
-  final void Function() onLongPressDown;
+  final void Function() onLongPressUp;
+  final void Function() onMoveStart;
   final void Function()? onTap;
 
   /// Pass tap event to the widgets below the mask.
@@ -222,15 +222,13 @@ class _CustomPopupMenuState extends State<CustomPopupMenu> {
   Widget build(BuildContext context) {
     var child = Material(
       child: GestureDetector(
-        onLongPressStart: (_) => widget.onLongPressDown(),
-        onLongPressUp: () {
-          if (!_controller!.menuIsShowing) {
-            widget.onLongPressUpWhilePopupHidden();
-          }
-        },
+        onLongPressUp: widget.onLongPressUp,
         onLongPressMoveUpdate: (details) {
+          if (_controller!.menuIsShowing) {
+            _controller!.hideMenu();
+            widget.onMoveStart();
+          }
           if (details.offsetFromOrigin.distance > 8) {
-            _controller?.hideMenu();
             widget.onLongPressMoveUpdate(details);
           }
         },
